perf/conntrack: fix flow configuration
Avoid the default action=NORMAL flow which can lead to inconsistencies
with the dpdk datapath. When the conntrack table is full, the ct()
actions will fail but will not cause the packet to be dropped. Instead,
it will be flagged as +trk+inv and passed on the next rule. Since the
last rule has no input filter, these packets are forwarded. Add an
explicit action=drop default rule at the end to avoid this.

NB: the kernel datapath seems to drop the packet immediately when any
ct() action fails.

Link: https://github.com/openvswitch/ovs/blob/v2.17.1/lib/conntrack.c#L1316=
Link: https://elixir.bootlin.com/linux/v5.18/source/net/openvswitch/conntrack.c#L1332
Signed-off-by: Robin Jarry <rjarry@redhat.com>