perf: add new create-vm role
Add support for creating a VM. This includes enabling libvirt,
downloading the base qcow2 image, customizing the image using virsh,
defining and starting the VM

The VM is configured with hugepages and pinned vCPUs on the same NUMA
socket for performance. testpmd is set to start automatically when the
vm starts through a systemd service.

testpmd is configured in mac forward mode with explicit peer mac
addresses on two vhost user interfaces.

Signed-off-by: Veda Barrenkala <vbarrenk@redhat.com>
Acked-by: Robin Jarry <rjarry@redhat.com>